Continue refactoring partitioned data reader
