+++ src/op_analytics/datapipeline/etl/intermediate/task.py
@@ -29,5 +29,5 @@ class IntermediateModelsTask:
     def __repr__(self):
         return (
             self.__class__.__name__
-            + f"[ctx: {self.data_reader.contextvars}, datset_paths: {self.data_reader.paths_summary}]"
+            + f"[ctx: {self.data_reader.partitions_dict()}, datset_paths: {self.data_reader.paths_summary}]"
         )
