+++ src/op_analytics/datapipeline/etl/loadbq/task.py
@@ -3,9 +3,8 @@ from dataclasses import dataclass
 from datetime import date
 
 from op_analytics.coreutils.logger import structlog
-from op_analytics.coreutils.partitioned import (
-    DataReader,
-)
+from op_analytics.coreutils.partitioned import DataReader
+from op_analytics.coreutils.time import date_fromstr
 
 
 log = structlog.get_logger()
@@ -33,19 +32,22 @@ def consolidate_chains(inputs: list[DataReader]) -> list[DateLoadTask]:
     """
     date_tasks: dict[date, DateLoadTask] = {}
     for inputdata in inputs:
-        if inputdata.dateval not in date_tasks:
-            date_tasks[inputdata.dateval] = DateLoadTask(
-                dateval=inputdata.dateval,
+        dateval = date_fromstr(inputdata.partition_value("dt"))
+        chain = inputdata.partition_value("chain")
+
+        if dateval not in date_tasks:
+            date_tasks[dateval] = DateLoadTask(
+                dateval=dateval,
                 dataset_paths=defaultdict(list),
                 chains_ready=set(),
                 chains_not_ready=set(),
             )
 
-        task = date_tasks[inputdata.dateval]
+        task = date_tasks[dateval]
         if not inputdata.inputs_ready:
-            task.chains_not_ready.add(inputdata.chain)
+            task.chains_not_ready.add(chain)
         else:
-            task.chains_ready.add(inputdata.chain)
+            task.chains_ready.add(chain)
 
         for dataset, paths in inputdata.dataset_paths.items():
             task.dataset_paths[dataset].extend(paths)
