+++ tests/op_datasets/etl/ingestion/test_reader.py
@@ -1,8 +1,9 @@
 import datetime
 
 import polars as pl
+
 from op_analytics.coreutils.partitioned import DataLocation
-from op_analytics.coreutils.partitioned.reader import are_inputs_ready
+from op_analytics.datapipeline.etl.ingestion.reader import are_inputs_ready
 
 MARKER_PATHS_DATA = [
     {
